use uint64_t instead of size_t for labeltype

+++ hnswlib/hnswlib.h
@@ -28,7 +28,7 @@
 #include <string.h>
 
 namespace hnswlib {
-    typedef size_t labeltype;
+    typedef 	uint64_t labeltype;
 
     template<typename T>
     static void writeBinaryPOD(std::ostream &out, const T &podRef) {
